pyproject.toml
[tool.poetry]
name = "djangoproject1"
version = "0.1.0"
description = ""
authors = ["woojin-an <dkswoojin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pymysql = "^1.1.1"
cryptography = "^43.0.1"
django-ninja = "^1.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
mypy = "^1.11.2"
isort = "^5.13.2"
django-stubs = "^5.0.4"
types-pymysql = "^1.1.0.20240524"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "djangoProject1.settings"